seperate video player component and show thumbnails
